### 2025-02-26 #7850 @xuanxu
@james-trayford nice to see STRAUSS submitted to JOSS! I will be the editor for this submission. I will be looking for reviewers this week. Please let me know if you can recommend reviewers.
