### 2025-02-26 #7850 @xuanxu
@johicasado :wave: would you be willing to review this paper for JOSS? I think your experience with sonification and Python makes you a great fit for this submission.
