### 2025-02-27 #7850 @warrickball
Just so everyone is aware, note that this paper has been submitted as a joint submission to American Astronomical Society (AAS) Journals, through which JOSS receives a small contribution from AAS Journals for our operating costs.  You can read more about this arrangement in the [announcement blog post](https://blog.joss.theoj.org/2018/12/a-new-collaboration-with-aas-publishing).
