### 2025-03-05 #7875 @editorialbot
[REVIEW]: STRAUSS: Sonification Tools & Resources for Analysis Using Sound Synthesis
**Submitting author:** @james-trayford (<a href="http://orcid.org/0000-0003-1530-1634">J. W. Trayford</a>)
**Repository:** https://github.com/james-trayford/strauss
**Branch with paper.md** (empty if default branch): v1.0.0
**Version:** v1.0.1_joss
