### 2025-03-17 522ef38
remove deprecated display import style, now removed from IPython
