### 2025-03-18 1d125bc
avoid strauss crashing when picking up hidden files in sample directories (e.g. .DS_store on mac), by forcing wav or WAV extension
