+++ src/strauss/generator.py
@@ -693,7 +693,7 @@ class Sampler(Generator):
                     logger.disabled = False
                     
             else:
-                wavs = sorted(Path(sampfiles).glob("*"))
+                wavs = sorted(Path(sampfiles).glob("{*.wav,*.WAV}"))
                 self.sampdict = {}
                 for w in wavs:
                     filename = Path(w).name
