+++ setup.cfg
@@ -1,6 +1,6 @@
 [metadata]
 name = strauss
-version = 1.0.0
+version = 1.0.1
 author = James Trayford
 author_email = james.trayford@port.ac.uk
 description = Sonification Tools and Resources for Astronomers Using Sound Synthesis
