+++ src/strauss/__init__.py
@@ -13,4 +13,4 @@ from . import sources
 from . import stream
 from . import presets
 
-__version__ = "0.4.0"
+__version__ = "1.0.1"
