+++ src/strauss/generator.py
@@ -694,7 +694,6 @@ class Sampler(Generator):
                     
             else:
                 wavs = sorted(Path(sampfiles).glob("*.[wW][aA][vV]"))
-                print(wavs)
                 self.sampdict = {}
                 for w in wavs:
                     filename = Path(w).name
