### 2025-03-19 #7850 @augustfly
@warrickball @james-trayford This is tagged as "AAS" -- can you let me know if a manuscript has been submitted to the AAS Journals or when it might be? I want to add this to the tracking spreadsheet that @arfon and I have been keeping but I can't find the matching AAS submission to log. Thanks!
