### 2025-03-19 #7850 @warrickball
@augustfly — Good catch. I applied the AAS label but I'm not quite sure why. The submission notes don't mention an AAS Journals submission, so I think I misinterpreted the AAS sidebar in the JOSS template as a sign that it's an AAS submission.

@james-trayford — If you can confirm there's no joint AAS submission, I'll remove the label (here and in the review thread).
