### 2025-03-21 #7850 @james-trayford
Hi both, apologies for the delay - you're right there is no intended AAS component to this - I was confused by the original AAS tagging but think i misread  @warrickball's explanation that this was being automatically pushed onto some joint AAS-JOSS scheme (which I was fine with) - feel free to take off any tags.
