### 2025-04-03 10151a2
Add animation script
