+++ src/strauss/animation.py
@@ -1,0 +1,777 @@
+""" The :obj:`animation` submodule: creates an audio-visual sequence including options for animating plots, mapping variables to selected sound properties. It also offers options to include audio captions, static slides and background videos.
+
+"""
+
+import glob
+import numpy as np
+import warnings
+import os
+import matplotlib.pyplot as plt
+from IPython.display import Audio
+from TTS.api import TTS
+from pathlib import Path
+import uuid
+import subprocess as sp
+import contextlib
+import io
+import sys
+from scipy.io import wavfile
+from strauss.utilities import resample
+import wavio as wav
+
+# ----------------------------------------------------------------------
+# Globals
+# ----------------------------------------------------------------------
+
+# fix audio sample rate
+SAMPRATE = 48000
+
+# maximum absolute sample value for audio peak normalisation
+MAXSAMP = (2**31)-1
+
+# supported sequence types
+seq_types = ['animation',
+             'image',
+             'text',
+             'blank',
+             'clip']
+
+defaults = {'fps': '30', 		# frames per second
+            'crf': '18',		# ffmpeg quality level
+            'invert_colours': True,	# dark on light is true
+            'transition_time': '30',	# sequence transition time [frames]
+            'breathing_time': '6',	# minimum time between separate sounds [frames]
+            #'dimensions': '4200x2100',
+            'background_video': './example_media/starfield.mov',
+            #'background_video': '/Users/jamestrayford/Downloads/stockvideo_01171.mov',
+            'dimensions': '3840x2160',   # video dimensions (4k by standard)
+            'orientation': 'vertical',
+            'transition_type': 'fade',
+            'slide_min_margin': '300',
+            'slide_key_black': '0',
+            'clip_override_duration': '1'
+            }
+
+# ----------------------------------------------------------------------
+# Useful dictionaries
+# ----------------------------------------------------------------------
+
+orient = {'vertical': '3840x2160',
+          'horizontal': '2160x3840'}
+
+res = {'4k': (3840, 2160),
+       '1080p': (1920, 1080),
+       '720p' : (1280, 720)}
+
+# ----------------------------------------------------------------------
+# Classes
+# ----------------------------------------------------------------------
+
+class Animate:
+    
+    def __init__(self, topdir, pars={}):
+        isdir = glob.glob(topdir)
+        Path(topdir).mkdir(parents=True, exist_ok=True)
+        if isdir:
+            if glob.glob(f'{isdir[0]}/*'):
+                warnings.warn(f"{topdir} is not empty, instead name "\
+                              "an empty directory, or a new one.")
+        self.topdir = topdir
+
+        # handle parameters
+        self.pars = defaults.copy()
+        for k in pars.keys():
+            self.pars[k] = pars[k]
+        self.pars['spf'] = SAMPRATE/int(self.pars['fps'])
+        self.spf = SAMPRATE/int(self.pars['fps'])
+
+        if self.spf % 1:
+            Exception("non integer samples-per-frame value, please use a standard video"\
+                      "fps (30,25,20) and audio sample rate (48000, 44100)")
+        
+        self.sequences = {}
+        self.frames = {}
+        self.seqlist = []
+        self.seqdx = 0
+
+        # stereo audio ramps to prevent dropouts, 30ms hard coded
+        self.aramplen = int(0.03*SAMPRATE)
+        self.arampin = np.column_stack([np.linspace(0,1, self.aramplen)]*2)
+        self.arampout = self.arampin[::-1]
+
+        # audio padding for breathing time and transitions
+        self.apadbreath = np.column_stack([np.zeros(int(self.pars['spf'] * float(self.pars['breathing_time'])))]*2)
+        self.apadtrans = np.column_stack([np.zeros(int(self.pars['spf'] * float(self.pars['transition_time'])))]*2)
[…]
