### 2025-04-11 #7875 @manwithfeathers
@xuanxu I believe my review is finished - can you check / confirm this as it's my first time reviewing for JOSS or in this format. I've really enjoyed this and STRAUSS is excellent.
