### 2025-04-11 #7875 @johicasado
@xuanxu I believe my review checklist is also complete. Thank you very much for this opportunity! I thoroughly enjoyed the process of reviewing for JOSS and learning more about this excellent tool.

I recommend the paper for submission.
