### 2025-04-14 #7875 @xuanxu
@james-trayford  looks like we're close to being done here. You mention `numpy` and `scipy` in your paper but there are not referenced in the citations, can you add them?
 - [numpy](https://github.com/numpy/numpy/blob/main/CITATION.bib)
 - [scipy](https://scipy.org/citing-scipy/)
