### 2025-04-14 7ad595c
add numpy and scipy references
