### 2025-04-15 98a51b6
strip AAS affiliation
