### 2025-04-15 #7875 @james-trayford
⬆️
- [x] ORCIDs for all
- [x] strip AAS association
- [x] add `numpy` and `scipy` references
- [x]  incorporate one line commit from main ([`generator.py#L696`](https://github.com/james-trayford/strauss/blob/99c90980cc70c15a844f3eb641c2b9cf8a8ca647/src/strauss/generator.py#L696)) that corrects `.wav` sample reading
