### 2025-04-15 #7875 @james-trayford
> > Nice!, a small detail missing: can you update the title of the [Zenodo archive](https://zenodo.org/records/15222076) to match the paper's title?
>
> [@james-trayford](https://github.com/james-trayford) And you may also correct the affiliations:
>
> ![Image](https://github.com/user-attachments/assets/80004e42-8b35-45e2-a275-1a9c092cf8ca)

Good catches - updated
