### 2025-05-04 ba64a86
Initial editorial tweaks for JOSS paper
