+++ joss/paper.bib
@@ -16,7 +16,7 @@
 	journal={Data Sonification Archive},
 	author={Lenzi, Sara and Ciuccarelli, Paolo and Liu, Houjiang and Hua, Yuan and Zizzi, Nicole},
 	year={2021},
-	month={Jan}} 
+	month={Jan}}
 
 @article{Lindborg23,
 AUTHOR={Lindborg, PerMagnus  and Lenzi, Sara  and Chen, Manni },
@@ -43,8 +43,8 @@ ISSN={1664-1078}
             Harris, Charles R. and Archibald, Anne M. and
             Ribeiro, Ant{\^o}nio H. and Pedregosa, Fabian and
             {van Mulbregt}, Paul and {SciPy 1.0 Contributors}},
-  title   = {{{SciPy} 1.0: Fundamental Algorithms for Scientific
-            Computing in Python}},
+  title   = {{SciPy} 1.0: Fundamental Algorithms for Scientific
+            Computing in {Python}},
   journal = {Nature Methods},
   year    = {2020},
   volume  = {17},
@@ -111,7 +111,7 @@ archivePrefix = {arXiv},
 }
 
 @ARTICLE{Trayford23b,
-       author = {{Trayford}, James W. and {Harrison}, C.~M. and {Hinz}, R.~C. and {Kavanagh Blatt}, M. and {Dougherty}, S. and {Girdhar}, A.},
+       author = {{Trayford}, J. W. and {Harrison}, C.~M. and {Hinz}, R.~C. and {Kavanagh Blatt}, M. and {Dougherty}, S. and {Girdhar}, A.},
         title = "{Inspecting spectra with sound: proof-of-concept and extension to datacubes}",
       journal = {RAS Techniques and Instruments},
      keywords = {sonification, integral field spectroscopy, galaxies, data methods, Astrophysics - Instrumentation and Methods for Astrophysics, Astrophysics - Astrophysics of Galaxies},
@@ -129,28 +129,42 @@ archivePrefix = {arXiv},
 }
 
 @inproceedings{Trayford23,
-author = {Trayford, James and Harrison, Chris},
+author = {Trayford, J. W. and Harrison, C.},
 year = {2023},
 month = {06},
 pages = {249-256},
-title = {Introducing strauss: A Flexible Sonification Python Package},
+title = {Introducing strauss: a flexible sonification {Python} package},
+booktitle = "{Proceedings of the 28th International Conference on Auditory Display (ICAD 2023)}",
+editor = {{Weger}, Marian and {Ziemer}, Tim and {R{\"o}nnberg}, Niklas},
 doi = {10.21785/icad2023.1978}
 }
 
 @ARTICLE{Trayford24,
-       author = {{Trayford}, J.~W. and {Youles}, S. and {Harrison}, C.~M. and {Bonne}, N.},
-        title = "{Ear to the Sky: Astronomical Sonification for Accessible Outreach, Education and Research with STRAUSS}",
-      journal = {\rmxaa},
+       author = {{Trayford}, J.~W. and {Youles}, S. and {Bonne}, N. and {Harrison}, C.~M.},
+        title = {Ear to the Sky: Astronomical Sonification for Accessible Outreach, Education and Research with {STRAUSS}},
+      journal = {Revista Mexicana de Astronom{\'i}a y Astrof{\'i}sica Serie de Conferencias},
      keywords = {Sonification, Multimodality, Data Analysis},
          year = 2024,
         month = jun,
        volume = {57},
-        pages = {42-46}
+        pages = {42-46},
+          url = {https://www.astroscu.unam.mx/rmaa/RMxAC..57/PDF/RMxAC..57_jtrayford-X.pdf}
 }
 
 
 
[…]
