+++ joss/paper.md
@@ -43,7 +43,7 @@ _Sonification_, or conveying data using non-verbal audio, is a relatively niche
 but growing approach for presenting data across multiple specialist
 domains including  astronomy, climate science, and beyond [@Lenzi21;
 @Zanella22; @Lindborg23]. The
-`strauss` Python package aims to provide such a tool, which builds upon previous
+`strauss` Python package aims to provide such a tool that builds upon previous
 approaches to provide a powerful means to explore different ways of expressing
 data, with fine control over the output audio and its format.
 `strauss` is a free, open source (FOSS) Python package, designed to allow flexible and
@@ -54,7 +54,7 @@ The remit of `strauss` is broad; it is intended to be able to bridge between
 _ad-hoc_ solutions for sonifying very particular datasets, and highly technical compositional
 and sound-design tools that are not optimised for sonification, or may have a steep
 learning curve. The code offers a range of approaches to sonification for a variety of
-contexts (e.g. science education, science communication, technical data analysis, etc). 
+contexts (e.g. science education, science communication, technical data analysis, etc.).
 To this end, `strauss` is packaged with a number of examples of different sonification
 approaches, and preset configurations to support a _low-barrier, high-ceiling_ approach.
 `strauss` has been used to produce both educational resources [@Harrison22], and
@@ -65,8 +65,8 @@ analysis tools [@Trayford23b].
 Sonification has great potential as a fundamental approach for interfacing with data.
 This provides new perspectives on data that complement visual approaches, as
 well as an accessible channel to data for those who cannot access visual
-presentation, for example those who are blind or visually impaired
-(BVI). As with the dominant approach of data 
+presentation, e.g. those who are blind or visually impaired
+(BVI). As with the dominant approach of data
 _visualisation_, what can constitute sonification is very broad, with different considerations for aspects
 such as the audience, information content and aesthetics of the sonification. In
 order for sonification to become more established and realise its potential as
@@ -75,17 +75,17 @@ sonification intuitive and accessible to those who routinely work with data.
 
 Unlike data _visualisation_, however, sonification of data is a far less developed
 methodology, with a lack of widely adopted, cross-domain tools and interfaces for those dealing with data
-to use. For example, in the Python programming language, a number of packages
-exist for visualising data, such as `matplotlib` [@Hunter07], `yt` [@Turk11],
-`seaborn` [@Waskom21], or `plotly` [@plotly]. The lack of dedicated and accessible
+to use. E.g., in the Python programming language, a number of packages
+exist for visualising data, such as Matplotlib [@Hunter07], yt [@Turk11],
+seaborn [@Waskom21], or Plotly [@plotly]. The lack of dedicated and accessible
 tools for sonifying data is a barrier to exploring the approach. Most solutions
 are piecemeal, using _ad-hoc_ tools to parse data and map it to properties of sound as
-well as generating and outputting sound in different formats. What's more, many 
+well as generating and outputting sound in different formats. What's more, many
 of the tools being repurposed to sonify data are proprietary and platform-dependent, requiring
 paid-for licenses.
 
-A number of effective python packages for data sonification have emerged e.g. `astronify`
-[@Brasseur23] and `SonoUno` [@Casado24]. However these are typically feature-limited
+A number of effective python packages for data sonification have emerged, e.g. `astronify`
+[@Brasseur23] and `SonoUno` [@Casado24], but these are typically feature-limited
 in how sonification is produced, namely continuous pitch-mapped sonification.
 
 Realising the potential of sonification may require a "crowdsourced" approach; via
@@ -93,14 +93,14 @@ broad adoption of the technique and innovation in sonification approaches driven
 the scientific community. In this context, we identify a need for Python package that:
 
[…]
