### 2025-05-04 #7875 @warrickball
I've just opened a [pull request](https://github.com/james-trayford/strauss/pull/44) to propose some editorial changes. If you're happy with those (and possibly tweak the title), I'll do another test run of the publication pipeline.
