### 2025-05-06 15f31b7
reformat STRAUSS->strauss in title for consistency, accompanying zenodo metadata change
