+++ joss/paper.md
@@ -1,5 +1,5 @@
 ---
-title: 'STRAUSS: Sonification Tools \& Resources for Analysis Using Sound Synthesis'
+title: 'strauss: Sonification Tools \& Resources for Analysis Using Sound Synthesis'
 tags:
   - Python
   - sonification
