### 2025-05-06 #7875 @james-trayford
@warrickball excellent, I've merged the https://github.com/james-trayford/strauss/pull/44 changes and also updated the title & accompanying [`zenodo` archive metadata](https://zenodo.org/records/15222076)
