### 2025-05-06 #7875 @editorialbot
:wave: @openjournals/aass-eics, this paper is ready to be accepted and published.
