### 2025-05-06 #7875 @james-trayford
@manwithfeathers @johicasado @xuanxu @warrickball thanks so much for all your efforts getting `strauss` published in JOSS! I appreciate you trying out the code and looking forward to building on it further from here. Keep in touch!
