### 2025-05-07 #7875 @warrickball
Hi @james-trayford, sorry to iterate here but the "&" in the paper title has rendered as "&amp;" on the JOSS page for the paper. It looks like the best fix is to replace "&" with "and" in the paper title. Could you make that change and let me know when you've done so? I can then re-run the publication pipeline, which should fix the website. Other metadata (notably the DOI) should remain unchanged.
