### 2025-05-07 #7875 @warrickball
That has indeed fixed it, so I can say congratulations @james-trayford, your paper has been published in JOSS! 🎉
